+++ changelog.md
@@ -3,3 +3,5 @@
 ## v0.3.0
 
 - add `selectableText` field in `tableStyleSpecs` config option
+- export `cssRulesFromSpecs` and `defaultTableStylesSpecs`
+- add FAQ to readme